add project method
* moved classes Segment, Point, World, Camera and Screen into Util.java
* fixed gradle run task
* fixed paths to textures
