+++ app/build.gradle
@@ -28,7 +28,7 @@ dependencies {
 
 application {
     // Define the main class for the application.
-    mainClass = 'App'
+    mainClass = 'Start'
 }
 
 tasks.named('test') {
